i_var.c: use unsigned int instead of size_t

+++ gschem/src/iface/i_vars.c
@@ -225,7 +225,7 @@ i_var_restore_group_color(EdaConfig *cfg, const char *group, const char *key,
   GError   *err = NULL;
   GdkColor *color;
   int      *array;
-  size_t    ridiculous = 4;
+  unsigned  ridiculous = 4;
 
   array = eda_config_get_int_list (cfg, group, key, &ridiculous, &err);
   if (err != NULL) {
